+++ src/main/java/br/com/meli/dhprojetointegrador/repository/SectionRepository.java
@@ -1,10 +1,9 @@
 package br.com.meli.dhprojetointegrador.repository;
 
+import br.com.meli.dhprojetointegrador.entity.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.meli.dhprojetointegrador.entity.Section;
-
 @Repository
-public interface SectionRepository extends JpaRepository<Section, Integer> {
+public interface SectionRepository extends JpaRepository<Section, Long> {
 }
